+++ src/main/java/cy/jdkdigital/trophymanager/compat/CobblemonCompat.java
@@ -1,0 +1,15 @@
+package cy.jdkdigital.trophymanager.compat;
+
+import com.cobblemon.mod.common.api.pokemon.PokemonProperties;
+import cy.jdkdigital.trophymanager.TrophyManager;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.level.Level;
+
+public class CobblemonCompat
+{
+    public static Entity create(Level level, CompoundTag tag) {
+        var properties = PokemonProperties.Companion.parse("species=\"" + tag.getCompound("Pokemon").getString("Species") + "\" level=" + level);
+        return properties.createEntity(level);
+    }
+}