+++ src/main/java/cy/jdkdigital/trophymanager/network/PacketOpenGui.java
@@ -1,11 +1,9 @@
 package cy.jdkdigital.trophymanager.network;
 
-import com.mojang.serialization.Codec;
 import cy.jdkdigital.trophymanager.TrophyManager;
 import cy.jdkdigital.trophymanager.client.gui.TrophyScreen;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.core.BlockPos;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;