+++ changelog.md
@@ -1,9 +1,65 @@
-1.21.0-2.1.9
+1.21.1-2.3.0
+
+- Proper Cobblemon support
+
+
+1.21.1-2.2.9
+
+- Added support for Ice & Fire entity variants
+- Added support for Starbuncle variants
+- Added support for Oh The Biomes We've Gone Wreath, Oddion and Pumpkin Warden
+
+
+1.21.1-2.2.8
+
+- Tropical fish variants
+
+
+1.21.1-2.2.7
+
+- Support for baby variants of zombie, husk, drowned, zombie villager and zombified piglins
+- Support for fox variants
+- Support for Living Things mob variants
+- Support for Pet Endermen (Enderman Overhaul)
+- Charged Creeper support
+
+
+1.21.1-2.2.6
+
+- Added trader llama variant
+- Added correct NBT and default properties for Man O War (BWG)
+
+
+1.21.1-2.2.5
+
+- Added cache nuking on level unload
+
+
+1.21.1-2.2.4
+
+- Fixed some rendering inconsistencies
+- Fixed productive bees compat
+
+
+1.21.0-2.2.1
+
+- Turn ender dragon upside down again
+
+
+1.21.1-2.2.0
+
+- Moved NBT map config to a datamap
+- Added datamap for default trophy settings like scale, offset and rotation
+- Added drop rate datamap where individual entity trophy drop rates can be configured
+- Fixed placing of ghost block when opening the gui while holding a block
+
+
+1.21.1-2.1.9
 
 - Fix crash from placing "empty" trophies
 
 
-1.21.0-2.1.8
+1.21.1-2.1.8
 
 - Fixed crash in 1.21.1 when putting trophy in curio slot